Add first script 3mf_bom_parser.py
Add 3mf_bom_parser.py this program takes a bambu labs 3mf file and
parses it to extract all the component STL files that were used in this
assembly. The original intent of this was to make it easy to generate a
BOM for multiboard assemblies so that someone else can reproduce the
work without violating the multboard license.

As for license - who knows - I would put it under the public domain but
I created it as an experiment in using Gemini's AI to generate code, and
to see how well I could construct prompts that guided it toward doing
what I want. I would assume that it is derived from other people's
copywritten work or at the very least the AI model was trained on that
work but the current law regarding copyright and the creation of
programs using AI is currently not well understood.

I will say however, that I am a fan of the GPL and would prefer if
anyone has any additions or bug fixes to this script they contribute
them back to this project rather than forking the code.

My creative contribution to this project was the idea for the work and
the 40 or so prompts which led Gemini to creating what I wanted.

Signed-off-by: Ben Woodard <woodard@redhat.com>